+++ Tweaks/UiAdjustment/RemoveQuestMarkerLimit.cs
@@ -1,4 +1,4 @@
-﻿using Dalamud.Utility.Signatures;
+using Dalamud.Utility.Signatures;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.System.String;
 using FFXIVClientStructs.FFXIV.Client.Game.UI;
@@ -10,7 +10,7 @@ namespace SimpleTweaksPlugin.Tweaks.UiAdjustment;
 
 [TweakName("Remove Quest Marker Limit")]
 [TweakDescription("Allow the map and minimap to display markers for more than 5 active quests.")]
-[Changelog(UnreleasedVersion, "Fix tweak not working since API 9")]
+[Changelog("1.9.1.1", "Fix tweak not working since API 9")]
 public unsafe class RemoveQuestMarkerLimit : UiAdjustments.SubTweak {
     private delegate void* SetQuestMarkerInfoDelegate(Map* map, uint index, ushort questId, Utf8String* name, ushort recommendedLevel);
     [TweakHook, Signature("E8 ?? ?? ?? ?? 0F B6 5B 0A", DetourName = nameof(SetQuestMarkerInfoDetour))]